fix explode with empty lists
