+++ polars/polars-core/src/series/any_value.rs
@@ -145,3 +145,42 @@ impl Series {
         }
     }
 }
+
+impl<'a> From<&AnyValue<'a>> for DataType {
+    fn from(val: &AnyValue<'a>) -> Self {
+        use AnyValue::*;
+        match val {
+            Null => DataType::Null,
+            Boolean(_) => DataType::Boolean,
+            Utf8(_) => DataType::Utf8,
+            Utf8Owned(_) => DataType::Utf8,
+            UInt32(_) => DataType::UInt32,
+            UInt64(_) => DataType::UInt64,
+            Int32(_) => DataType::Int32,
+            Int64(_) => DataType::Int64,
+            Float32(_) => DataType::Float32,
+            Float64(_) => DataType::Float64,
+            #[cfg(feature = "dtype-date")]
+            Date(_) => DataType::Date,
+            #[cfg(feature = "dtype-datetime")]
+            Datetime(_, tu, tz) => DataType::Datetime(*tu, (*tz).clone()),
+            #[cfg(feature = "dtype-time")]
+            Time(_) => DataType::Time,
+            List(s) => DataType::List(Box::new(s.dtype().clone())),
+            #[cfg(feature = "dtype-struct")]
+            StructOwned(payload) => DataType::Struct(payload.1.to_vec()),
+            #[cfg(feature = "dtype-struct")]
+            Struct(_, fields) => DataType::Struct(fields.to_vec()),
+            #[cfg(feature = "dtype-duration")]
+            Duration(_, tu) => DataType::Duration(*tu),
+            UInt8(_) => DataType::UInt8,
+            UInt16(_) => DataType::UInt16,
+            Int8(_) => DataType::Int8,
+            Int16(_) => DataType::Int16,
+            #[cfg(feature = "dtype-categorical")]
+            Categorical(_, rev_map) => DataType::Categorical(Some(Arc::new((*rev_map).clone()))),
+            #[cfg(feature = "object")]
+            Object(o) => DataType::Object(o.type_name()),
+        }
+    }
+}
